:sparkles: 32 feature add inactive attribute to seatingtype and equipment (#35)
* added isActive to entities and db-scripts

* added isActive to dtos and regenerated api

* added isActive to entities and db-scripts

* added isActive to dtos and regenerated api

* added isActive to Entity, dtos and adjusted mapper

* changed delete behavior to check for active state and introduced new exception with new msg

* updated api

* updated the edit form for seatingType and equipment

* updated table via adding new column for status

* changed variant to text and added disabled prop

* added scoped attributes to itemActions slot

* customized actions to be disabled upon active items

* typed set the itemAction slot and generic slot and form slot

* added mapping for toEntity

* exchanged custom card with card-form

* :white_check_mark: fixed test compilation error

* :truck: renamed migration script to match the order

* :art: moved template content to prop

* :rotating_light: spotless formated

* :rotating_light: fixed pmd

+++ raumreservierung-backend/src/main/java/de/muenchen/raumreservierung/equipment/EquipmentService.java
@@ -1,7 +1,9 @@
 package de.muenchen.raumreservierung.equipment;
 
+import static de.muenchen.raumreservierung.common.ExceptionMessageConstants.MSG_CANNOT_DELETE_ACTIVE;
 import static de.muenchen.raumreservierung.common.ExceptionMessageConstants.MSG_NOT_FOUND;
 
+import de.muenchen.raumreservierung.common.ConflictException;
 import de.muenchen.raumreservierung.common.NotFoundException;
 import de.muenchen.raumreservierung.security.Authorities;
 import java.util.List;
@@ -32,8 +34,7 @@ public class EquipmentService {
 
     @PreAuthorize(Authorities.EQUIPMENT_MANAGE)
     public Equipment updateEquipment(final Equipment equipment, final UUID equipmentId) {
-        final Equipment foundEquipment = equipmentRepository.findById(equipmentId)
-                .orElseThrow(() -> new NotFoundException(String.format(MSG_NOT_FOUND, equipmentId)));
+        final Equipment foundEquipment = getEntityOrThrowException(equipmentId);
         foundEquipment.updateFrom(equipment);
         log.debug("Updating equipment to {}", foundEquipment);
         return equipmentRepository.save(foundEquipment);
@@ -41,8 +42,19 @@ public class EquipmentService {
 
     @PreAuthorize(Authorities.EQUIPMENT_MANAGE)
     public void deleteEquipment(final UUID equipmentId) {
+        final Equipment toDelete = getEntityOrThrowException(equipmentId);
+
+        if (toDelete.isActive()) {
+            throw new ConflictException(String.format(MSG_CANNOT_DELETE_ACTIVE, equipmentId));
+        }
+
         log.debug("Deleting equipment {}", equipmentId);
         equipmentRepository.deleteById(equipmentId);
     }
 
+    private Equipment getEntityOrThrowException(final UUID equipmentId) {
+        return equipmentRepository
+                .findById(equipmentId)
+                .orElseThrow(() -> new NotFoundException(String.format(MSG_NOT_FOUND, equipmentId)));
+    }
 }
